Initial commit.
This code is not completely functional. It might not been integrated
into the continuous integration system because of an incompatibility
between com.btmatthews.maven.plugins.inmemdb:inmemdb-maven-plugin and
Maven 3.1. It would still be possible to do such integration by
commenting the lines for pre-integration testing in privacy/POM.xml.

Thales plans to replace the existing persistence based on the file
system by a persistence based on a database.